do summary math with unsigned longs, helps avoid overflow

+++ main.c
@@ -83,10 +83,10 @@ int main(int argc, char* argv[])
     qsort(timings, total, sizeof(unsigned long), cmp_ul_asc);
 
     printf("Request time (ms)\n");
-    printf(" 50%%: %d\n", (int)(timings[total / 2] / 1000.0));
-    printf(" 75%%: %d\n", (int)(timings[(int)(total * 0.75)] / 1000.0));
-    printf(" 95%%: %d\n", (int)(timings[(int)(total * 0.95)] / 1000.0));
-    printf(" max: %d\n", (int)(timings[total - 1] / 1000.0));
+    printf(" 50%%: %lu\n", (unsigned long)(timings[total / 2] / 1000.0));
+    printf(" 75%%: %lu\n", (unsigned long)(timings[(unsigned long)(total * 0.75)] / 1000.0));
+    printf(" 95%%: %lu\n", (unsigned long)(timings[(unsigned long)(total * 0.95)] / 1000.0));
+    printf(" max: %lu\n", (unsigned long)(timings[total - 1] / 1000.0));
     printf("\n");
 
     printf("Failures: %lu\n", failures);
